pyproject.toml
[project]
name = "agent-handson"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agent-framework-core>=1.0.0b251104",
    "agent-framework>=1.0.0b251104",
    "azure-cognitiveservices-speech>=1.47.0",
    "azure-identity>=1.25.1",
    "python-dotenv>=1.2.1",
    "packaging>=25.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[tool.uv]
prerelease = "allow"
